src code for exercies 1 and 4.
In exercise 1, I do not understand what meansa list of edges because when we talk about graph, we often talk about adjacency list or adjacency matrix. Well, that why I do something strange

The solution for exercise 5 was implemented in Eulerian example. To find the number of connected components if the graph is not connected, just do a bit modification. So I will not do this :3

For exercise 2 and 3, I don understand much but any way, if I understand it right, it is easy